# Fix handwritten cosine similarity and add test for NaNs in AEVComputer