@IgnacioJPickering 2020-12-17
@yueyericardo How does the cuda test work? It is randomly failing for me