@yueyericardo 2020-12-18
> @yueyericardo How does the cuda test work? It is randomly failing for me

not sure, looks a rerun passed